+++ packages/babel-plugin/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @kuma-ui/babel-plugin
 
+## 0.1.0
+
+### Minor Changes
+
+- Overriding styles defined within tagged template literals using style props
+
+### Patch Changes
+
+- Updated dependencies
+  - @kuma-ui/sheet@0.0.4
+  - @kuma-ui/system@0.1.2
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @kuma-ui/sheet@0.0.3
+  - @kuma-ui/system@0.1.1
+
 ## 0.0.3
 
 ### Patch Changes