+++ packages/core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @kuma-ui/core
 
+## 0.0.4
+
+### Patch Changes
+
+- Overriding styles defined within tagged template literals using style props
+
+## 0.0.3
+
+### Patch Changes
+
+- Replace global whitespace removal regex with more specific regex for CSS properties #23
+  Refactor HTML Tags. #24
+  Fix incorrect setup example for Next.js #25
+  fix position 'bottom, inset' css prop values #26
+
 ## 0.0.2
 
 ### Patch Changes