+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @kuma-ui/vite
 
+## 0.1.2
+
+### Patch Changes
+
+- Overriding styles defined within tagged template literals using style props
+- Updated dependencies
+  - @kuma-ui/babel-plugin@0.1.0
+  - @kuma-ui/sheet@0.0.4
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @kuma-ui/sheet@0.0.3
+  - @kuma-ui/babel-plugin@0.0.4
+
 ## 0.1.0
 
 ### Minor Changes